requirements.txt
neo4j>=5.0
langchain-neo4j>=0.1
langchain-ollama>=0.1
chromadb>=0.5
requests>=2.31
python-dotenv>=1.0
